+++ tokenizers/src/models/wordlevel/mod.rs
@@ -169,8 +169,8 @@ impl Model for WordLevel {
         self.vocab.get(token).copied()
     }
 
-    fn id_to_token(&self, id: u32) -> Option<String> {
-        self.vocab_r.get(&id).cloned()
+    fn id_to_token(&self, id: u32) -> Option<&str> {
+        self.vocab_r.get(&id).map(String::as_ref)
     }
 
     fn get_vocab(&self) -> &HashMap<String, u32> {